@@ -4,12 +4,12 @@ import 'consts.dart';
 import 'errors/validation_error.dart';
 
 abstract class Entity extends Equatable {
-  final String id;
+  final String? id;
 
   Entity(this.id);
 
   @override
-  List<Object> get props => [this.id];
+  List<Object?> get props => [this.id];
 
   /// Validate the entity
   /// @return all errors from validating, empty list if there are no validation errors
@@ -21,7 +21,7 @@ abstract class Entity extends Equatable {
     return errors;
   }
 
-  static validateId(final String id, final List<ValidationInfo> errors) {
+  static validateId(final String? id, final List<ValidationInfo> errors) {
     if (id == null) {
       errors.add(new ValidationInfo(type: ValidationTypes.idNotDefined));
     } else if (id == '') {
@@ -29,7 +29,7 @@ abstract class Entity extends Equatable {
     }
   }
 
-  static validateName(final String name, final List<ValidationInfo> errors) {
+  static validateName(final String? name, final List<ValidationInfo> errors) {
     if (name == null || name.length < ValidationConsts.nameLengthMin) {
       errors.add(new ValidationInfo(
         type: ValidationTypes.nameTooShort,
