@@ -1,5 +1,3 @@
-import 'package:meta/meta.dart';
-
 import '../../../../core/entity.dart';
 import '../../../../core/errors/validation_error.dart';
 
@@ -10,16 +8,16 @@ class Practice extends Entity {
   final PracticeMistakes memoryFlubs;
 
   Practice({
-    String id,
-    @required this.pieceId,
-    @required this.date,
-    @required this.technicalMistakes,
-    @required this.memoryFlubs,
+    String? id,
+    required this.pieceId,
+    required this.date,
+    required this.technicalMistakes,
+    required this.memoryFlubs,
   }) : super(id);
 
   @override
-  List<Object> get props {
-    final List<Object> props = super.props;
+  List<Object?> get props {
+    final List<Object?> props = super.props;
 
     props.add([
       this.pieceId,
@@ -47,11 +45,11 @@ class Practice extends Entity {
   }
 
   Practice copy({
-    String id,
-    String pieceId,
-    DateTime date,
-    PracticeMistakes technicalMistakes,
-    PracticeMistakes memoryFlubs,
+    String? id,
+    String? pieceId,
+    DateTime? date,
+    PracticeMistakes? technicalMistakes,
+    PracticeMistakes? memoryFlubs,
   }) {
     return new Practice(
       id: id != null ? id : this.id,
