@@ -2,6 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:faker/faker.dart';
 import 'package:mockito/mockito.dart';
+import 'package:my_musical_repertoire/core/errors/error.dart';
 import 'package:my_musical_repertoire/features/piece/domain/repositories/piece_repository.dart';
 import 'package:my_musical_repertoire/features/piece/domain/usecases/piece_remove.dart';
 
@@ -11,8 +12,8 @@ class _MockPieceRepository extends Mock implements PieceRepository {}
 
 void main() {
   group("Remove Piece should (#usecase #cold) ->", () {
-    PieceRemove usecase;
-    _MockPieceRepository mockPieceRepository;
+    late PieceRemove usecase;
+    late _MockPieceRepository mockPieceRepository;
 
     setUp(() {
       mockPieceRepository = _MockPieceRepository();
@@ -21,9 +22,9 @@ void main() {
 
     test('Call the repository to remove the piece with the specified id', () async {
       final String id = _faker.guid.guid();
-      when(mockPieceRepository.removePiece(any)).thenAnswer((_) async => Right(id));
+      when(mockPieceRepository.removePiece(id)).thenAnswer((_) async => Right(id));
 
-      final result = await usecase(id);
+      final Either<Error, String> result = await usecase(id);
 
       expect(result, Right(id));
       verify(mockPieceRepository.removePiece(id));
